refactor: MainPage를 pages/의 최상위 경로로 이동
MainPage가 굳이 main/이라는 폴더 안에서 관리되어야 할 필요가 없는 것 같음.
